+++ libs/libft/src/math/ft_abs.c
@@ -6,20 +6,20 @@
 /*   By: mmichele <mmichele@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/30 21:40:16 by mmichele          #+#    #+#             */
-/*   Updated: 2025/07/10 22:56:40 by mmichele         ###   ########.fr       */
+/*   Updated: 2025/10/31 23:54:49 by mmichele         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-double	ft_abs(double value)
+float	ft_abs(float value)
 {
 	if (value >= 0)
 		return (value);
 	return (-value);
 }
 
-t_bool	fts_abs(double *value)
+t_bool	fts_abs(float *value)
 {
 	if (*value <= -FT_DBL_MAX)
 		return (0);
