+++ libs/libft/src/math/vector/vec3.c
@@ -6,14 +6,14 @@
 /*   By: mmichele <mmichele@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/10 22:31:24 by mmichele          #+#    #+#             */
-/*   Updated: 2025/07/11 23:34:08 by mmichele         ###   ########.fr       */
+/*   Updated: 2025/10/31 23:53:36 by mmichele         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
 /* Returns the multiplication of a vector by a scalar. */
-t_vec	vec_scal(const t_vec v, const double scalar)
+t_vec	vec_scal(const t_vec v, const float scalar)
 {
 	t_vec	res;
 
@@ -24,7 +24,7 @@ t_vec	vec_scal(const t_vec v, const double scalar)
 }
 
 /* Returns the division of a vector by a scalar. */
-t_vec	vec_rscal(const t_vec v, const double scalar)
+t_vec	vec_rscal(const t_vec v, const float scalar)
 {
 	t_vec	res;
 
@@ -35,7 +35,7 @@ t_vec	vec_rscal(const t_vec v, const double scalar)
 }
 
 /* Multiplies a vector by a scalar. */
-void	vec_iscal(t_vec *src, const double scalar)
+void	vec_iscal(t_vec *src, const float scalar)
 {
 	src->x *= scalar;
 	src->y *= scalar;
@@ -43,7 +43,7 @@ void	vec_iscal(t_vec *src, const double scalar)
 }
 
 /* Divides a vector by a scalar. */
-void	vec_irscal(t_vec *src, const double scalar)
+void	vec_irscal(t_vec *src, const float scalar)
 {
 	src->x /= scalar;
 	src->y /= scalar;
